🆙Update: 登録フォームの再描画ロジックを改善
