+++ app/views/templetes/_form.html.erb
@@ -1,47 +1,50 @@
-<%= form_with model: forms,
-              url: templetes_path,
-              method: :post,
-              data: { action: "turbo:submit-end->modal-frame#close" } do |f| %>
+<%= turbo_frame_tag "templete_form_frame" do %>
+  <div data-controller="templete-form">
 
-  <%# NOTE: save失敗時にレンダリングを制御するためindexのselect_tagの値を受け取る %>
-  <%= f.hidden_field :select_tag_value %>
+    <%= form_with model: forms,
+                  url: templetes_path,
+                  method: :post,
+                  data: { action: "turbo:submit-end->modal-frame#close" } do |f| %>
 
-  <!-- 保管場所名フィールド -->
-  <div class="mb-4">
-    <label for="input_location_name" class="modal-label">保管場所名</label>
-    <%= render 'location_fields', f: f, forms: forms, locations_name: locations_name %>
-    <span class="text-xs text-dull-red"><%= forms.errors[:location_name].first if forms.errors[:location_name].present? %></span>
-  </div>
+      <!-- 保管場所名フィールド -->
+      <div class="mb-4">
+        <label for="input_location_name" class="modal-label">保管場所名</label>
+        <%= render 'location_fields', f: f, forms: forms, locations_name: locations_name %>
+        <span class="text-xs text-dull-red"><%= forms.errors[:location_name].first if forms.errors[:location_name].present? %></span>
+      </div>
 
-  <!-- ストック名・状態・個数フィールド群 -->
-  <div class="w-full font-bold text-sm flex items-center gap-2 mb-1">
-    <div class="basis-1/2 ml-7">ストック名</div>
-    <div class="basis-1/3 text-center">状態・個数</div>
-  </div>
-  <div data-templete-select-target="container" class="w-full">
-    <% forms.stock_forms.each_with_index do |stock_el, index| %>
-      <%= render 'stock_fields', f: f, stock_el: stock_el, index: index %>
-    <% end %>
-  </div>
+      <!-- ストック名・状態・個数フィールド群 -->
+      <div class="w-full font-bold text-sm flex items-center gap-2 mb-1">
+        <div class="basis-1/2 ml-7">ストック名</div>
+        <div class="basis-1/3 text-center">状態・個数</div>
+      </div>
+      <div data-templete-form-target="container" class="w-full">
+        <% forms.stock_forms.each_with_index do |stock_el, index| %>
+          <%= render 'stock_fields', f: f, stock_el: stock_el, index: index %>
+        <% end %>
+      </div>
 
-  <%# NOTE: 追加フォーム用に忍ばせる隠しDOM %>
-  <template data-templete-select-target="templete">
-    <%= render 'stock_fields', f: f, stock_el: TempletesStockForm.new(model: 0, exist_quantity: 1), index: "__INDEX__" %>
-  </template>
+      <%# NOTE: 追加フォーム用に忍ばせる隠しDOM %>
+      <template data-templete-form-target="templete">
+        <%= render 'stock_fields', f: f, stock_el: TempletesStockForm.new(model: 0, exist_quantity: 1), index: "__INDEX__" %>
+      </template>
 
-  <div class="my-4">
-    <div class="flex items-center space-x-2">
-      <%= button_tag type: "button", data: { action: "click->templete-select#add" }, id: "add_form" do %>
-        <div class="border-2 border-dull-green text-dull-green rounded cursor-pointer flex justify-center items-center space-x-2 px-2 py-1">
-          <svg class="size-4 text-dull-green" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
-            <line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" />
-          </svg>
-          <span>追加</span>
+      <div class="my-4">
+        <div class="flex items-center space-x-2">
+          <%= button_tag type: "button", data: { action: "click->templete-form#add" }, id: "add_form" do %>
+            <div class="border-2 border-dull-green text-dull-green rounded cursor-pointer flex justify-center items-center space-x-2 px-2 py-1">
+              <svg class="size-4 text-dull-green" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+                <line x1="12" y1="5" x2="12" y2="19" /><line x1="5" y1="12" x2="19" y2="12" />
+              </svg>
+              <span>追加</span>
+            </div>
+          <% end %>
+          <span data-templete-form-target="notice" class="hidden text-sm text-dull-green">フォームは10個まで追加できます</span>
         </div>
-      <% end %>
-      <span data-templete-select-target="notice" class="hidden text-sm text-dull-green">フォームは10個まで追加できます</span>
-    </div>
-  </div>
+      </div>
 
-  <%= f.submit "保存", class: "modal-submit w-full" %>
+      <%= f.submit "保存", class: "modal-submit w-full" %>
+    <% end %>
+
+  </div>
 <% end %>
